requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.9
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.1.3
blinker==1.8.2
ccxt==4.4.63
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.2.1
cryptography==44.0.1
cycler==0.12.1
dill==0.4.0
Django==5.2
djangorestframework==3.16.0
dlib==19.24.4
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
filelock==3.16.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
idna==3.10
imageio==2.34.1
isort==6.0.1
itsdangerous==2.2.0
jax==0.4.37
jaxlib==0.4.36
Jinja2==3.1.4
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.15
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.7
namex==0.0.9
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.82
opt_einsum==3.4.0
optree==0.15.0
packaging==24.0
pandas==2.2.3
pillow==10.3.0
platformdirs==4.3.7
propcache==0.3.0
protobuf==4.25.5
pycares==4.5.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.8.0
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-image==0.23.2
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.30
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.5.0
tifffile==2024.5.22
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
Werkzeug==3.0.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
